==> 1 Extract JSONs.ipynb <==
# for basic array manipulation
import numpy as np

# for submodule dedicated to image processing (n-dimensional images)
from scipy import ndimage

# for printing  numpy image arrays as images
import matplotlib.pyplot as plt

from glob import glob

from skimage import color
from skimage import io

import pandas as pd

PATH = "ISIC Archival Images/"

def extractJSONs(path):
    dataset = glob(path+"ISIC_*.jpg")
    dataset.sort() # All the ISIC image files in the folder
    datasetJSON = glob(path+"*.json")
    datasetJSON.sort()
    
    
    indices, diagnoses = [], []
    for json in datasetJSON:
        lesionID = pd.read_json(json).iloc[0]["name"]
        indices.append(lesionID)
        diagnosis = pd.read_json(json).iloc[1]["meta"]["diagnosis"]
        diagnoses.append(diagnosis)
    lesionTargets = pd.DataFrame(diagnoses, index = indices)
    
    # get folder-of-origin for later search expediency
    lesionTargets["origin"] = path
    
    return lesionTargets

from tqdm import tqdm_notebook

setPaths = ["MSK-1/", "MSK-2/","MSK-3/","MSK-4/","MSK-5/","SONIC/","UDA-1/","UDA-2/"]
sets = []

for path in tqdm_notebook(setPaths):
    sets.append(extractJSONs(PATH+path))

ARCHIVE = pd.concat(sets)
ARCHIVE = pd.get_dummies(ARCHIVE, columns=[0])
ARCHIVE

ARCHIVE.columns = ["origin", "AKIEC","BCC","DF","MEL","NV","BKL"]
ARCHIVE["VASC"] = 0
ARCHIVE.index.name = "image"
len(ARCHIVE)

ARCHIVE.to_csv("ARCHIVE_Task3_Training_GroundTruth.csv")

ARCHIVE.sum()

pd.read_csv("ARCHIVE_Task3_Training_GroundTruth.csv")
