==> 2 Partition Sets.ipynb <==
import pandas as pd
import numpy as np

HAM = pd.read_csv("ISIC2018_Task3_Training_GroundTruth/ISIC2018_Task3_Training_GroundTruth.csv", index_col=0)
ARCHIVE = pd.read_csv("ARCHIVE_Task3_Training_GroundTruth.csv", index_col=0)
ARCHIVE = ARCHIVE[["origin","MEL","NV","BCC","AKIEC","BKL","DF","VASC"]]

HAM["origin"] = "ISIC2018_Task3_Training_Input/"

ARCHIVE.sum()

len(pd.concat([HAM,ARCHIVE]))==(len(HAM)+len(ARCHIVE))
WHOLE = pd.concat([HAM,ARCHIVE], sort=False)



stratifier = np.array(HAM)

from sklearn.model_selection import train_test_split
HAM_X_train, HAM_X_test = train_test_split(HAM.index,test_size=0.2, stratify=stratifier)

TEST = HAM.T[HAM_X_test].T

TRAIN = pd.concat([HAM.T[HAM_X_train].T,ARCHIVE], sort=False)

TEST.shape, TRAIN.shape

TEST.to_csv("Principal: Processed Images/TEST.csv")
TRAIN.to_csv("Principal: Processed Images/TRAIN.csv")
