==> 2.5 Remove Classes.ipynb <==
import pandas as pd
import numpy as np
import cv2
from skimage import color
from skimage import io

from scipy import misc

TEST = pd.read_csv("Principal: Processed Images/TEST.csv", index_col=0)
TRAIN = pd.read_csv("Principal: Processed Images/TRAIN.csv", index_col=0)

TRAIN.sum()

TRAIN.columns

TRAIN = TRAIN.drop(TRAIN[(TRAIN.BCC == 1) | (TRAIN.AKIEC == 1) | (TRAIN.DF == 1) | (TRAIN.VASC == 1)].index)
TEST = TEST.drop(TEST[(TEST.BCC == 1) | (TEST.AKIEC == 1) | (TEST.DF == 1) | (TEST.VASC == 1)].index)

TRAIN = TRAIN.drop(["BCC", "AKIEC", "DF", "VASC"], axis = 1)
TEST = TEST.drop(["BCC", "AKIEC", "DF", "VASC"], axis = 1)

TEST.sum()

TEST.to_csv("Principal: Processed Images/TESTmod.csv")
TRAIN.to_csv("Principal: Processed Images/TRAINmod.csv")
