==> 5 Prep Format.ipynb <==
currently notebook configured to not use any data augmentation

import pandas as pd
import numpy as np
from glob import glob

TRAINAUG = pd.read_csv("Processed Images/NEWTRAIN/TRAINAUG.csv", index_col=0).drop("origin", axis=1)
TRAIN = pd.read_csv("Processed Images/TRAINmod.csv", index_col=0)#.drop("origin", axis=1)
TEST = pd.read_csv("Processed Images/TESTmod.csv", index_col=0).drop("origin", axis=1)

TRAIN = TRAIN[TRAIN.origin=="ISIC2018_Task3_Training_Input/"].drop("origin", axis=1)

ALLTRAIN = pd.concat([TRAINAUG, TRAIN], sort=False)
#ALLTRAIN = TRAIN

ALLTRAIN.shape

ALLTRAIN.sum()

# numpy arrays for input and targets
imageList = []
targetList = []

imageValList = []
targetValList = []



imageListPlain = []

from tqdm import tqdm_notebook
from skimage import io

Fill in imageList and targetList with 4 folds

#checkDirectories = ["FOLD2","FOLD3","FOLD4", "FOLD5"] # here exclude fold that you are training on
counter1 = 0
counter2 = 0


filelist = glob('Principal: Processed Images/NEWTRAIN/'+direc+'AUG/ISIC_*.jpg')
for name in tqdm_notebook(filelist):
    if(name[35:47] in ALLTRAIN.index):
        imageList.append(io.imread(name))
        targetList.append(np.array(ALLTRAIN.loc[name[35:47]]))
        counter2+=1

Fill in imageValList and targetValList with remaining fold

filelist = glob('Processed Images/NEWTRAIN/FOLD1/ISIC_*.jpg')
for name in tqdm_notebook(filelist):
    if(name[32:44] in ALLTRAIN.index):
        imageValList.append(io.imread(name))
        targetValList.append(np.array(ALLTRAIN.loc[name[32:44]]))

imageList = np.array(imageList)
targetList = np.array(targetList)

imageList.shape, targetList.shape#, imageValList.shape, targetValList.shape

Compute weights for CNNs:

(weight of class) = (total number of samples)/(number of samples in class)

np.sum(targetList, axis = 0)/np.sum(targetList)

np.sum(targetValList, axis = 0)/np.sum(targetValList)

targetList

np.repeat(np.sum(targetList), 3)/np.sum(targetList, axis=0)

np.repeat(np.sum(targetValList), 3)/np.sum(targetValList, axis=0)

#np.save("targetList", targetList)
#np.save("imageValList", imageValList)
#np.save("targetValList", targetValList)

targetValList.shape

np.mean(imageList, axis = tuple(range(imageList.ndim-1)))

np.mean(imageValList, axis = tuple(range(imageValList.ndim-1)))

imageListPlain = np.array(imageListPlain)
np.mean(imageListPlain, axis = tuple(range(imageListPlain.ndim-1)))

## Prep TEST Folder

testList = []
targetTestList = []

TEST.sum()

filelist = glob('Processed Images/TEST/ISIC_*.jpg')
for name in tqdm_notebook(filelist):
    testList.append(io.imread(name))
    targetTestList.append(np.array(TEST.loc[name[22:34]])) 

testList = np.array(testList)
targetTestList = np.array(targetTestList)

np.mean(testList, axis = tuple(range(testList.ndim-1)))

testList.shape, targetTestList.shape

np.sum(targetTestList, axis = 0)/np.sum(targetTestList)

np.repeat(np.sum(targetTestList), 3)/np.sum(targetTestList, axis=0)

np.savez_compressed("", testList=testList, targetTestList=targetTestList)


## Check HAM mean values

veryPlain = []

# keep origin
TRAIN = pd.read_csv("Processed Images/TRAIN.csv", index_col=0)

checkDirectories = ["FOLD2","FOLD3","FOLD4", "FOLD5"] # here exclude fold that you are training on
for direc in tqdm_notebook(checkDirectories):
    filelist = glob('Processed Images/NEWTRAIN/'+direc+'/ISIC_*.jpg')
    for name in tqdm_notebook(filelist):
        if(TRAIN.loc[name[32:44]]["origin"]=="ISIC2018_Task3_Training_Input/"):
            veryPlain.append(io.imread(name))

veryPlain = np.array(veryPlain)

np.mean(veryPlain, axis = tuple(range(veryPlain.ndim-1)))
